--- src/app/pages/not-found/not-found.component.html
@@ -1,9 +1,0 @@
-<div fxLayout="column">
-  <mat-card fx-Flex style="margin-top: 10%; width: 50%; margin-left: 25%;" class="mat-elevation-z8">
-    <mat-card-title style="text-align: center; font-size: 48px" class="mat-headline">404 - Page Not Found</mat-card-title>
-      <br><br>
-      <mat-card-actions>
-        <button fxFlex mat-raised-button color="primary" [routerLink]="['/']">Return to Homepage</button>
-      </mat-card-actions>
-  </mat-card>
-</div>
