+++ src/app/shared/guard/authGuard.ts
@@ -27,7 +27,7 @@ export class AuthGuard implements CanActivate {
     if (isAuthenticated) {
       return true;
     } else {
-      this.router.navigate(['/session/login']);
+      this.router.navigate(['/session/unauthorized']);
       return false;
     }
   }
