+++ src/app/shared/navigation/navigation.component.html
@@ -1,0 +1,36 @@
+<!--Toolbar html-->
+<mat-toolbar class="menu" role="header" color="primary">
+  <mat-toolbar-row>
+    <button routerLink="/" mat-button class="toolbar__icon-button mat-button">
+      <span style="margin-left:5px !important">Bobs Repair Shop</span>
+    </button>
+    <button mat-button class="mat-button" [routerLink]="['/about-us']">About</button>
+    <button mat-button class="mat-button" [routerLink]="['/contact-us']">Contact Us</button>
+    <button class="toolbar_icon_button" [matMenuTriggerFor]="menu" mat-icon-button>
+      <mat-icon>account_circle</mat-icon>
+    </button>
+    <mat-menu #menu="matMenu">
+      <button routerLink="/session/login" mat-menu-item>
+        <mat-icon> perm_identity</mat-icon>
+        <span>Login</span>
+      </button>
+      <button routerLink="/my-profile" mat-menu-item>
+        <mat-icon> perm_contact_calendar</mat-icon>
+        <span>My Profile</span>
+      </button>
+      <button mat-menu-item [matMenuTriggerFor]="admin">
+        <mat-icon>settings</mat-icon>
+        <span>Admin</span>
+      </button>
+        <mat-menu #admin="matMenu">
+          <button routerLink="/admin/users" mat-menu-item>User Management</button>
+          <button routerLink="/admin/security-questions" mat-menu-item>Security Questions</button>
+          <button routerLink="/admin/sales-chart" mat-menu-item>Sales Chart</button>
+        </mat-menu>
+      <button mat-menu-item>
+        <mat-icon>exit_to_app</mat-icon>
+        <span>LogOut</span>
+      </button>
+    </mat-menu>
+  </mat-toolbar-row>
+</mat-toolbar>
